Bug 6889 Mapping files into memory to reduce I/O.
When looking in *.def files for definitions, rather than opening when a
file is consulted, it is mapped entirely in memory using mmap(). A list
of structs holds a list of mapped files with the adresses of the mappings.

When a definition from a file is needed, the list is traversed to find
the filestart pointer for that file.  If it is not found, it is opened,
mapped, and added to the list.  Then the filestart pointer is returned.

The mapped file is parsed in almost the same way as previously.

A function has been created to unmap the files using munmap(). However,
it is not know where that function should be used.

This change is transparent to any caller of SeqUtil_parsedef(filename,key).