Changed the logic for copying links
When one of
	$src/listings
	$src/logs
	$src/sequencing
is a symbolic link, an empty directory is made with the same name in the
destination experiment.

Note that at line 112, the trailing slashes needed to be removed from the
exclude statements since a link named logs is not excluded by
	--exclude "logs/"
even if the link points to a directory. Directories are still correctly
excluded even without the trailing slash.