Bug5454B User's choices are realized by creating a script and calling it after the import
The script, ${TMPDIR}/post_import.sh does
   rm the dir that should be a link
   makes a link to the user specified path

** This is a combination of two commits
** First commit message

Import.tcl creates post_import scrip

The script does what it is supposed to do, but for some reason, a copy of the
link it makes ends up in the source experiment.  I.E Importing an exp and asking
for bin to point towards the source experiment works: the new imported exp's bin
folder is a link pointing to the original's bin folder, however the original's
bin folder contains a copy of this link.

This doesn't happen when I run the generated script from the terminal so I
suspect that what is happening is that the import_maestro_exp and the
post_install scripts are run at the same time and rsync does some stuff that I
don't understand.

** Second commit message

Calling the post_import script in a place in the TCL code where I know it wil
run after import_maestro_exp has finished

This way, it gets called after import_maestro_exp completes.