Bug7050 Added support for long options
 --  This is a combination of 5 commits.
 --  The first commit's message is:
Added AIX-Linux support for long options.

The definition of long options is given in struct option long_opts.

The correspondence between long option and short option is given in the
long_opt[] array in the 4th attribute of each option.  This 4th attribute is
what is returned by getopt_long() when it parses that long option.  Therefore
the switch/case statement remains unchanged.

Since long options are not natively supported on AIX, the files getopt.h and
getopt_long.c are included with the project.

When they are no longer needed (if we stop using AIX), then for all the *_main.c
files, we can
a) change
   #include "getopt.h"
to
   #include <getopt.h>,
b)remove the makefile target
   getopt_long.o : getopt_long.c getopt.h
c) for each of the *_main.c files, remove
   getopt_long.o
from the corresponding makefile target's dependency list and from the arguments
of the compiler call in that target.

In case we want to revert back to short options, just change every call to
getopt_long to getopt and remove the last two arguments, and remove the struct
option long_opts. Also do the makefile changes b) and c) and get rid of the
getopt.h and getopt_long.c files, and remove the #include "getopt.h" from the
*_main.c files.  Or just revert this commit, but if reverting causes conflicts
or other confusing git things, undoing the changes manually may take less time
than figuring out confusing things.

 --  This is the 2nd commit message:

Updated documentation

 --  This is the 3rd commit message:

more doc

 --  This is the 4th commit message:

DOC

 --  This is the 5th commit message:

Finished new documentation