Inserted behavior of doesNodeExist() into getFlowInfo()
The function now has the behavior of doesNodeExist if the argument _nodeDataPtr
is NULL.  In that case, the places where the existence of the node specified by
_nodePath can be confirmed or infirmed, the function returns 0 if the node
exists and 0 if it doesn't.

Also, in that case, some of the work of getFlowInfo is skipped because it is
unncessary and can lead to errors.