Completing bug6486 (removing syscalls)
When trying to get submitDependencies to work without the system calls, I
backtracked and undid some of my previous work in order to redo it step by step
and figure out where the problem came from.  When I finally did get it to work,
I forgot to go back and re-introduce the part about emitting a nodelogger
message and the part about checking to not submit nodes twice.