Changed the test in processContainerEnd and processContainerBegin
The goal of the test was to determine whether the current node is an L or an
L(i) by checking if it's loop_args has an attribute whose name matches the nodes
name.

To be more robust against bad names, the test also checks to make sure that the
node for which we are asking the question is a loop, NpassTasks or switch.