When a submission fails, execution is not interrupted
Reintroduced system calls in submitDependencies so that an exit statement in the
submission of a dependency does not stop our own execution

Note the use of #define SEQ_USE_SYSTEM_CALLS_FOR_DEPS. we would prefer to use
function calls rather than system calls so if we do decide to do that in the
future, it will be possible to switch easily.  The code is already there and is
equivalent except for the fact that the system call will contain any exit
statements encountered while submitting dependencies.