Bug7049 completed forgotten part (bug7232)
SEQ_EXP_HOME was stored in a static global array in maestro.c and I missed two
places where it was still used.  It is now completely gone.

This also pertains to bug7232 because there should have been a lookup of
SEQ_DATE in the environment if the datestamp is not supplied with "-d".