SeqLoops_getLoopContainerExtensionsInReverse() now works with wildcards dependencies in switches
Two things needed to be solved for this to work.

First of all, the preceding commit (Rewriting SeqLoops_getLoopContainerExtensionsInReverse()
hereafter referred to as to SLGLCEIR()) dealt with atoi() being called on switch
values.

The second issue was when the dependency was on the actual switch or loop:
<DEPENDS_ON dep_name="thing/switch_a/loop_b/loop_c" index="switch_a=00,loop_b=*,loop_c=*/>.
We construct a SeqNodeDataPtr for the node dep_name, then for each name-value
pair in index, we look in _nodeDataPtr->loops for a LoopsPtr whose name field
matches the name in the name-value pair.  However, loop_c will not be in there
because _nodeDataPtr->loops is only for container loops.

getFlowInfo() does put the informations with the nodeDataPtr, but it puts it in
_nodeDataPtr->data for numerical loops, and doesn't for switches.

The changes made to fix this were to first make getFlowInfo put the switch value
in _nodeDataPtr->data, and second, to make SLGLCEIR() look for information in
the _nodeDataPtr->data when no matching loop could be found in the
_nodeDataPtr->loops list.