Changed to more standard style for nodeinfo's filters
This is the style used by some libraries like OpenGL for sending a variable
number of boolean variables to a function.

Instead of setting global ints, we set individual bits in an unsigned int.

   filters |= NI_SHOW_ALL  (where NI_SHOW_ALL = (1 << N) )

will set the Nth bit.

We check the boolean variable by doing

   if( filters & NI_SHOW_ALL )

and we can call nodeinfo with the filters of our choice:

   nodeinfo( ..., NI_RESOURCE|NI_SHOW_DEP, ...).

From the command line, the filters start as a comma separated string and the
wrapper function in nodeinfo_main.c parses this string into the unsigned int.

The filters that nodeinfo receives can be passed to it's subroutines like
getFlowInfo, and one can add filters if one so wishes.

   nodeinfo(..., filters, ...){
      ...
      getFlowInfo(..., filters|NI_SHOW_DEP);

or remove them

      getFlowInfo(..., filters & ~NI_SHOW_DEP);
      ...
   }

   filters |= (1 << N) sets the Nth bit
   filters &= ~(1 << N) unsets the Nth bit.

Another option would have been a struct of ints, but this has the disadvantage
that you couldn't write

   nodeinfo(..., NI_SHOW_ALL|NI_SHOW_DEP,...);

You would have to create the struct and set it's attributes, then pass the
struct.

   struct filters ;
   initFilters(&filters);
   filters->show_all = 1;
   filters->show_dep = 1;
   nodeinfo(..., filters,...);

Note that an initialiser like

   struct filters = {0,0,1,1,0,0};

is less recommended because of maintainability.  If we were to add an attribute
to this struct, we might have to update many initialisers to make sure that the
new field is 0-initialized.

Also, note that this is less error-prone: I created a macro for each filter that
was searched for when parsing the string. The filter "type" was not one of them.

the call

   nodeinfo(..., "type",...)

worked fine, but the call

   nodeinfo(..., NI_TYPE,...)

caused a compilation error because NI_TYPE was an undeclared symbol whereas
"type" is just a string like any other.

Now, nodeinfo_main has an "if, else if, else if" structure with and "else
raiseError()" added at the end:

   if( strcmp(token,"all") == 0 ) filters |= NI_SHOW_ALL;
   else if (strcmp...
   ...
   else raiseError("Unrecognized token");

which would have caught the bad token, BUT ONLY AT RUNTIME.  Whereas with the
current setup, errors like this or like spelling "Type" instead of "type" get
caught at compile time.