Created the SeqNodeCensus utility.
This provides the interface given by SeqNodeCensus.h whose main function is

   getNodeList(seq_exp_home);

It will return a list of all the nodes in the experiment.  Right now, it returns
the list of all the nodes of the experiment (regardless of the datestamp*).  In
order to be able to do something with that list, say calling nodeinfo, we must
not be affected by switch branches.  Therefore, since nodeinfo crashes when it
tries to go into a switch branch that 'does not exist with the current
datestamp' the list also contains a switch_args field.  Nodeinfo passes that
field to getFlowInfo() which puts it in it's FlowVisitor->switch_args field to
be used at the appropriate points.

NOTE:
   1) The switch_args field of the PathArgNode is either empty string or not
      empty.

   2) The constructor for the FlowVisitor sets fv->switch_args to NULL if the
      recieved switch_args is empty string.

   3) If getFlowInfo() receives NULL for it's switch_args argument, behavior is
      identical to before.

   4) In the futur this switch_args becomes unnecessary, but the functionnality
      will be left in because it may be useful to be able to specify ourselves
      what switch_branch to take.

* That means that we get all nodes for all switch branches.

ORIGINAL COMMIT MESSAGES (Leaving them in but I don't think they're useful to read)

* This is a combination of 13 commits.
* The first commit's message is:
================================
changed to a list pair

* This is the 2nd commit message:
=================================

switch_branches are now in a separate string.

* This is the 3rd commit message:
=================================

added extra argument to nodeinfo for use in creating the nodeinfo_database.

This switch_arg string will be stored in the flow visitor so that the
Flow_findSwitchItem() can use it to find the correct switch item regarless of
the context (datestamp, day of week etc.

* This is the 4th commit message:
=================================

Initial setup for making switch args available to the Flow_findSwitchItem() function.

* This is the 5th commit message:
=================================

using switch_arg to correctly enter switch_items

I need to do some extra work, because the switch_args, as they are created,
contain all the values of one SWITCH_ITEM.

That is if <SWITCH_ITEM name="1,2,3">, then we will have switch_args like
switch1=1,2,3,switch_2=00.  ANd the commas will screw up my parsing.

* This is the 6th commit message:
=================================

Everything works and doing some cleanup, splitting into functions and making headers

* This is the 7th commit message:
=================================

Removed paranoia initializations

* This is the 8th commit message:
=================================

Cleanup and commenting.  Plus I moved the saving and restoring of node into the main internal routine.  I didn't feel like making another commit.

* This is the 9th commit message:
=================================

Fixed memory leak

* This is the 10th commit message:
==================================

Moved node census functions into new file with it's makefile target.

* This is the 11th commit message:
==================================

Removed pointer type warnings that I ignored during development

* This is the 12th commit message:
==================================

Added CFLAGS to compilation rules for object files

* This is the 13th commit message:
==================================

Cleanup and function headers.  Node census looks finished.