Changed SeqLoops_indexToExt for better maintainability.
The initial version did the least amount of operations and was much faster, but
slightly hurt maintainability in the case where we decide to change how extensions
and indices work.  I left it there protected by a macro and with some comments.

It's a tradeoff in favor of maintainability over performance.