Added datestamp dependance
This changes the SeqNodeCensus utility somewhat:

   getNodeList(seq_exp_home, datestamp)

now takes a datestamp argument and returns a list of all the nodes that exist
at the current datestamp.

I did leave the option to not specify a datestamp and have it behave exactly
as before in case that could be useful in the futur (I'm always thinking
about the futur).

At the same time, when tsvinfo iterates through the list and calls
getNodeResources(nodeDataPtr,...) it sets the datestamp in the nodeDataPtr so
that the correct resources are read into it to be put in the keyed_list.

I wanted dates to be padded in tsvinfo_main, so:

To reuse date padding code, I added a function in SeqUtil:

   SeqUtil_addPadding(char *dst, char *src, char c, int length)

and a macro in SeqDatesUtil.h : PADDED_DATE_LENGTH and replaced the code for
padding in tictac_main.c by a call to this function, and used this function
in tsvinfo_main.c to padd the date with '0's.

** At the same time as I did this, I made keynames lowercase, and changed their
names to fit with xflow's naming scheme.

This is a combination of 5 commits.
===================================

The first commit's message is:
==============================

Modifications while working with xflow

Made leaf keys lowercase
Added a couple keys

Modified the test file accordingly

This is the 2nd commit message:
===============================

Added some comments, and changed Flow_findSwitchType to use xmlGetProp

This is the 3rd commit message:
===============================

Made tsvinfo datestamp dependant

This is the 4th commit message:
===============================

Added a utility function for doing padding.

This is the 5th commit message:
===============================

Finishing touches for datestamp dependance of tsvinfo