Fix for default values of loop data.
To save xflow from having to do checks, and so that default values can reside in
one place, I added default values in SeqLoopsUtil.h in the form of macros.

One problem remains: parseLoopAttributes is called only for loops on the
container path of a node. Not for the node itself.  That is not the problem, the
problem is that when getting loop information for the node itself, that is done
with parseNodeSpecifics, a function that doens't know about loops.

Therefore, with it's code, the default values cannot be set there.  So, to have
default values in a single location, they have to be set in tsvinfo.

The correction to this problem would involve changing the way loops work.  That
shoudl be done the same way as what is being done for dependencies: since the
attributes of a loop are all known at compile time, it makes more sense to use a
struct than a NameValuesPtr.  Even if for one particular loop, the fields of the
struct are not all used, the benefits of using a struct versus a NameValuesPtr
are substantial.  And aside from having to modify the code, there are NO
downsides.