Going with OPTION 1 for printing loops keyed list
To see the different options, look at the code right before this commit.  I had
several ideas:

1) Only keys that have a value exist. So non-loops don't have a .loop keyed
   list.  And loops only have the subkeys that exist, so regular loops don't
   have loop.expression key, and expression loops don't have loop.start loop.end
   etc.  keys.

2) Every node has a .loop key, but only the subkeys that have a value exist. For
   loops, this is the same as option 1, but for non-loops, they will have the
   .loop key with no subkeys.

3) Every key is there, even the ones without values.  Regular loops, the value
   loop.expression will be "(null)", and for expression loops, .loop.start will
   be "(null)". Also, non-loops will have a .loop keyed list with the five
   subkeys, each with "(null)" value.  This option is ridiculous.

4) Only loops have a .loop key, but with every subkey.  So regular loops would
   have the .loop.expression key, and it's value would be "(null)".

Option 1 was chosen because one can check for the presence of the key with
TsvInfo_haskey command in TCL.

One can also get values for keys that may not exist with the TsvInfo_getInfoPlus
function.

And for keys that should exist, one should use TsvInfo_getNodeInfo function
because it will halt the program if the key doesn't exist, allowing errors to be
caught and reported in the speediest manner.