Static functions should not be declared in header files ...
They need to be declared before you use them, so put the declarations at the top of the .c file.

In the absence of -Werror=implicit-function-declaration, this is still necessary, because if
there is no static declaration, and the function is used, say my_undeclared_function, then the
implicit function definition rule will declare a non-static function, and when the compiler
reaches the implementation, it will complain that it doesn't matche the declaration (you won't
find that declaration, the error message note will point you to the first use of that function
which will be confusing.

This confusion can be avoided by using -Werror=implicit-function-declaration.