Fix extraneous declaration of MLLServerConnectionFid ...
Again, it's not super clear, MLLServerConnectionFid exists in only one place:
in SeqUtilServer.c.  To be able to use the variable in OthreFile.c, that file
must have a declaration 'extern int MLLServerConnectionFid;'.  This declaration
is added to .c files that want to access the global int by including the header
file that corresponds to the .c file where the actual int is created.